+++ test/testThreads1.cc
@@ -46,7 +46,7 @@ int main()
             std::vector<std::thread> threads;
             int j1=-thr_evts[i].at(e).vertices().size();
             int j2=thr_evts[i].at(e).particles().size();
-            int d=(j2-j1)/NmaxThreads;
+            int d=((long)(j2)-(long)(j1))/NmaxThreads;
             std::vector<int> ids;
             ids.push_back(0);
             for (int j=j1; j<j2; j+=d)
